+++ front/src/api/postApi.js
@@ -1,20 +1,6 @@
-import axios from "axios";
+// API 미구축 - 참고용 mock
+// 서버 연동 시: axios/fetch로 GET /posts 호출
 
-//const API = axios.create({
-//  baseURL: "http://YOUR_DJANGO_SERVER/api",
-//});
-
-//export const getPosts = async () => {
-//  const res = await API.get("/posts");
-//  return res.data;
-//};
-
-//export const getPostDetail = async (id) => {
-//  const res = await API.get(`/posts/${id}`);
-//  return res.data;
-//};
-
-//export const createPost = async (data) => {
-//  const res = await API.post("/posts", data);
-//  return res.data;
-//};
+export const getPosts = async () => {
+  return { results: [] };
+};
