+++ front/src/components/map/MapPostPreviewCard.js
