--- front/src/src/api/postApi.js
@@ -1,6 +1,0 @@
-// API 미구축 - 참고용 mock
-// 서버 연동 시: axios/fetch로 GET /posts 호출
-
-export const getPosts = async () => {
-  return { results: [] };
-};
