--- front/src/src/constants/category.js
@@ -1,9 +1,0 @@
-// ui_doc 기준: 스터디/운동/맛집/취미/네트워킹/기타
-export const CATEGORIES = [
-  "스터디",
-  "운동",
-  "맛집",
-  "취미",
-  "네트워킹",
-  "기타",
-];
