+++ front/src/styles/simpleScreenStyles.js
